GPTVoiceAssistant/requirements.txt
audio_recorder_streamlit==0.0.8
boto3==1.24.28
numpy==1.23.5
openai==0.26.4
openai_whisper==20230124
pandas==1.4.4
Pillow==9.4.0
playsound==1.3.0
pydub==0.25.1
python-dotenv==0.21.1
requests==2.28.1
scipy==1.9.3
sounddevice==0.4.5
streamlit==1.12.0
whisper==1.1.10
